+++ SadnaSrc/SadnaSrc/StoreCenter/Discount.cs
@@ -29,7 +29,7 @@ namespace SadnaSrc.StoreCenter
 
         }
         //assume that the User discountCode is the right one, the check is done by the Store itself
-        public double CalcDiscount(int basePrice)
+        public double CalcDiscount(double basePrice)
         {
             if (Percentages)
                 return basePrice * (1 - ((double)DiscountAmount/100));
