+++ SadnaSrc/SadnaSrc/StoreCenter/StoreShoppingService.cs
@@ -163,7 +163,7 @@ namespace SadnaSrc.StoreCenter
                 {
                     if (stockListItem.Discount.discountType == discountTypeEnum.Visible)
                         if (stockListItem.Discount.checkTime())
-                        stockListItem.Product.BasePrice = (int)stockListItem.Discount.CalcDiscount(stockListItem.Product.BasePrice);
+                        stockListItem.Product.BasePrice = stockListItem.Discount.CalcDiscount(stockListItem.Product.BasePrice);
                 }
                 _shopper.AddToCart(stockListItem.Product, store,quantity);
                 MarketLog.Log("StoreCenter", "add product successeded");
